Chore organize firebase utility functions
