+++ src/components/organisms/AddNewPost/index.js
@@ -2,10 +2,10 @@ import React, {useReducer, useEffect} from 'react'
 import PropTypes from 'prop-types'
 import {Redirect} from 'react-router-dom'
 
-import {getUserTimelines} from '../../../utilities/query'
-import {uploadMediaToStorage} from '../../../utilities/write'
+import {getUserTimelines} from '../../../utilities/firebase/query'
+import {uploadMediaToStorage} from '../../../utilities/firebase/write'
 import resizeImage from '../../../utilities/jimp/image-manipulation'
-import saveNewPost from '../../../utilities/saveNewPost'
+import saveNewPost from '../../../utilities/firebase/saveNewPost'
 import CheckboxInput from '../../atoms/CheckboxInput'
 import DatePickerInput from '../../atoms/DatePickerInput'
 import MediaUpload from '../../molecules/MediaUpload'
