+++ src/utilities/firebase/delete.js
@@ -1,4 +1,4 @@
-import firebase from '../firebase'
+import firebase from '../../firebase'
 
 /**
  * Delete 'posts/postKey' from '/posts'.
