+++ src/utilities/firebase/generateFirebaseUpdateObject.js
