+++ src/utilities/firebase/query.js
@@ -1,4 +1,4 @@
-import firebase from '../firebase'
+import firebase from '../../firebase'
 
 /**
  * Get all timelines from user.
