+++ src/utilities/firebase/updatePost.js
@@ -1,4 +1,4 @@
-import firebase from '../firebase'
+import firebase from '../../firebase'
 import generateFirebaseUpdateObject from './generateFirebaseUpdateObject'
 
 /**
